WIP: use Media Source Extension
Currently uses a video element to decode the video on the sender.
Uses MediaStream & MediaRecorder to blobs from the video and then uses
Media Source Extension to play the video on the receiver.

The video element could be replaced with a custom canvas renderer, or
the MediaStream/MediaRecorder part could be replaced with ffmpeg.wasm.

There is much work remaining, such as improving the UI, supporting
subtitles, throttling render, showing video length on receiver, handling
play/pause commands, and adding back WebTorrent support. The current
implementation also has a tendency to crash.